requirements.txt
aiogram~=2.23.1
Pillow~=9.3.0
requests~=2.28.1
loguru~=0.6.0
pyppeteer~=1.0.2
APScheduler~=3.9.1.post1
aiosqlite~=0.17.0
pydantic~=1.10.2
beautifulsoup4~=4.11.1
aiohttp~=3.8.3
psycopg2~=2.9.5
asyncpg~=0.27.0
pytest~=7.2.0
pdf2image~=1.16.2
pre-commit
